+++ CHANGELOG.md
@@ -10,6 +10,9 @@
     * new file view for all embed types
     * image thumb stays in preivew mask, also stay in fulllscreen mask
     * in fullscreen only pinch (mobile 2 touches) and zoom (mousewheel) possible, pan not possible yet (move)
+* add a font loader
+* update thirdparty list
+* add about us to help route
 
 #########
 # RC 0.5.0 - 12.01.2015